Refactor: Create separate functional CartView component
