+++ client/components/OrderView.js
@@ -2,6 +2,7 @@ import React from 'react'
 import axios from 'axios'
 import {withRouter} from 'react-router-dom'
 import {connect} from 'react-redux'
+import {CartView} from './index'
 
 class OrderView extends React.Component {
   constructor() {
@@ -17,7 +18,6 @@ class OrderView extends React.Component {
       `/api/orders/${this.props.match.params.orderId}`
     )
     const order = response.data
-    console.log(order)
     this.setState(prevState => ({
       ...prevState,
       order
@@ -28,7 +28,15 @@ class OrderView extends React.Component {
     if (!this.state.order.id || this.props.products.length === 0) {
       return <div>...loading</div>
     }
-    const {firstName, lastName, address, contents, email, id} = this.state.order
+    const {
+      firstName,
+      lastName,
+      address,
+      contents,
+      email,
+      id,
+      totalPrice
+    } = this.state.order
     const products = this.props.products
     return (
       <div>
@@ -38,14 +46,11 @@ class OrderView extends React.Component {
         <div>Address: {address}</div>
         <div>Email: {email}</div>
         <div>
-          {contents.map(item => {
-            const product = products.find(e => e.id === item.productId)
-            return (
-              <div key={product.id}>{`Name: ${product.name}, Quantity: ${
-                item.quantity
-              }`}</div>
-            )
-          })}
+          <CartView
+            contents={contents}
+            products={products}
+            totalPrice={totalPrice}
+          />
         </div>
       </div>
     )
