+++ client/components/index.js
@@ -14,4 +14,5 @@ export {default as AllProducts} from './AllProducts'
 export {default as SingleCategory} from './SingleCategory'
 export {default as CheckoutForm} from './CheckoutForm'
 export {default as OrderView} from './OrderView'
+export {default as CartView} from './CartView'
 export {Login, Signup} from './AuthForm'
